Update the Navbar
