+++ config/initializers/cors.rb
@@ -1,4 +1,4 @@
-# config/initializers/cors.rb
+config/initializers/cors.rb
 Rails.application.config.middleware.insert_before 0, Rack::Cors do
   allow do
     origins 'http://localhost:3000'
